YASK: Adjust the use of explicit dimension sizes on operator invocation
The current scheme uses dimension sizes in terms of absolute indices on
the time axis. This means that `t=2` does not run two timesteps, but
instead it runs a single timestep up to the second entry in time.
Assuming that the first entry is the initial condition this will thus
perform only a single timestep calculation.